put prompts in own file, fixed logic in orgOptions
