+++ index.js
@@ -2,69 +2,15 @@
 
 // dependencies
 const inquirer = require("inquirer");
-const viewQueries = require("../utils/viewQueries.js");
-const editQueries = require("../utils/editQueries.js");
+const viewQueries = require("./utils/viewQueries.js");
+const editQueries = require("./utils/editQueries.js");
 const {
-    isValidString,
-    isValidNumber,
-    isValidSalary,
-} = require("../utils/utilfunctions.js");
+    mainMenu,
+    viewOptions,
+    editOptions,
+} = require("./utils/inquirerQuestions.js");
 
 // prompt user for input - should be recursive
-const mainMenu = [
-    {
-        type: "list",
-        name: "menu",
-        message: "Select an action: ",
-        choices: ["View Organization", "Edit Organization", "Exit"],
-    },
-];
-
-const viewOptions = [
-    {
-        type: "list",
-        name: "menu",
-        message: "Select an action: ",
-        choices: [
-            new inquirer.Separator(" = View All = "),
-            "View all departments",
-            "View all roles",
-            "View all employees",
-            new inquirer.Separator(" = View Filtered = "),
-            "View employees by department",
-            "View employees by role",
-            "View employees by manager",
-            "View total budget by department",
-            new inquirer.Separator(" = Exit = "),
-            "Return to main menu",
-            "Exit application",
-        ],
-    },
-];
-
-const editOptions = [
-    {
-        type: "list",
-        name: "menu",
-        message: "Select an action: ",
-        choices: [
-            new inquirer.Separator(" = Add = "),
-            "Add a department",
-            "Add a role",
-            "Add an employee",
-            new inquirer.Separator(" = Update = "),
-            "Update employee",
-            "Update role",
-            new inquirer.Separator(" = Remove = "),
-            "Remove department",
-            "Remove role",
-            "Remove employee",
-            new inquirer.Separator(" = Exit = "),
-            "Return to main menu",
-            "Exit application",
-        ],
-    },
-];
 
 // REQUIRED
 // A view all departments prints a table to console with department names and ids
@@ -87,7 +33,7 @@ function init() {
 
 function orgOptions(prompt) {
     let view = "View Organization";
-    let questions = prompt && view ? viewOptions : editOptions;
+    let questions = prompt === view ? viewOptions : editOptions;
     inquirer.prompt(questions).then((res) => {
         if (res.menu === "Exit application") {
             return;
@@ -98,27 +44,5 @@ function orgOptions(prompt) {
         }
     });
 }
-
-// function viewOrg() {
-//     inquirer.prompt(viewOptions).then((res) => {
-//         if (res.menu === "Exit application") {
-//             return;
-//         } else if (res.menu === "Return to main menu") {
-//             init();
-//         } else {
-//             viewQueries(res.menu);
-//         }
-//     });
-// }
-
-// function editOrg() {
-//     inquirer.prompt(editOptions).then((res) => {
-//         if (res.menu === "Exit application") {
-//             return;
-//         } else if (res.menu === "Return to main menu") {
-//             init();
-//         }
-//     });
-// }
 
 init();
