+++ utils/editQueries.js
@@ -3,17 +3,18 @@
 function editQueries(choice) {
     // switch case for determining which query to get for editing
     switch (choice) {
-        case value:
-            break;
-        case value:
-            break;
-        case value:
-            break;
-        case value:
-            break;
-        case value:
-            break;
+        // case value:
+        //     break;
+        // case value:
+        //     break;
+        // case value:
+        //     break;
+        // case value:
+        //     break;
+        // case value:
+        //     break;
         default:
+            console.log("Editing logic goes here")
             break;
     }
 }
