+++ utils/viewQueries.js
@@ -3,17 +3,18 @@
 function viewQueries(choice) {
     // switch case for calling which endpoint for viewing
     switch (choice) {
-        case value:
-            break;
-        case value:
-            break;
-        case value:
-            break;
-        case value:
-            break;
-        case value:
-            break;
+        // case value:
+        //     break;
+        // case value:
+        //     break;
+        // case value:
+        //     break;
+        // case value:
+        //     break;
+        // case value:
+        //     break;
         default:
+            console.log("Viewing logic goes here")
             break;
     }
 }
